requirements.txt
kaldifst>1.7.0
kaldilm
kaldialign
num2words
kaldi-decoder
sentencepiece>=0.1.96
pypinyin==0.50.0
tensorboard
typeguard
dill
omegaconf>=2.3.0
hydra-core>=1.3.2
onnx>=1.15.0
onnxruntime>=1.16.3
onnxoptimizer
onnxsim
onnxconverter_common

# style check session:
black==22.3.0
isort==5.10.1
flake8==5.0.4

# cantonese word segment support
pycantonese==3.4.0
packaging
